+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Voucher fields CSS avoid to be overriden by WooCommerce defaults
+* New setting for recipient heading text
 
 ### 1.3.0: September 14th, 2020
 * Fix version number